+++ centrifuge-app/package.json
@@ -32,7 +32,6 @@
     "@makerdao/multicall": "^0.12.0",
     "@polkadot/extension-dapp": "^0.44.1",
     "@polkadot/react-identicon": "^2.8.2",
-    "@safe-global/safe-apps-sdk": "^7.11.0",
     "@stablelib/blake2b": "^1.0.1",
     "@styled-system/css": "^5.1.5",
     "@styled-system/should-forward-prop": "^5.1.5",
