+++ centrifuge-react/src/components/WalletProvider/useConnectEagerly.ts
@@ -1,10 +1,8 @@
-import SafeAppsSDK from '@safe-global/safe-apps-sdk'
 import { getWalletBySource } from '@subwallet/wallet-connect/dotsama/wallets'
 import { Wallet } from '@subwallet/wallet-connect/types'
 import { GnosisSafe } from '@web3-react/gnosis-safe'
 import * as React from 'react'
 import { EvmConnectorMeta } from './evm/connectors'
-import { createConnector } from './evm/utils'
 import { Action, getPersisted } from './useWalletState'
 
 let triedEager = false
@@ -19,44 +17,28 @@ export function useConnectEagerly(
   async function tryReconnect() {
     try {
       setIsTrying(true)
-
       const { wallet: source, type } = getPersisted()
+      const isProbablyGnosis = window !== window.parent
 
-      const gnosisSdk = new SafeAppsSDK()
-      // if gnosis safe is not within the context, the getInfo call will never resolve so we need to timeout
-      // https://github.com/safe-global/safe-apps-sdk/issues/263#issuecomment-1029835840
-      const isGnosis = !!(await Promise.race([
-        gnosisSdk.safe.getInfo(),
-        new Promise<undefined>((resolve) => setTimeout(resolve, 200)),
-      ]))
-
-      if (isGnosis) {
-        const [gnosisSafe] = createConnector<GnosisSafe>((actions) => new GnosisSafe({ actions }))
-
-        const gnosisSafeConnector = {
-          id: 'gnosis-safe',
-          title: 'Gnosis Safe',
-          installUrl: '',
-          logo: {
-            src: '',
-            alt: 'Gnosis Safe',
-          },
-          connector: gnosisSafe,
-          get installed() {
-            return true
-          },
-          get shown() {
-            return false
-          },
+      if ((type === 'evm' && source) || isProbablyGnosis) {
+        let wallet
+        if (isProbablyGnosis) {
+          wallet = evmConnectors.find((c) => c.connector instanceof GnosisSafe)
+        } else {
+          wallet = evmConnectors.find((c) => c.id === source)
         }
 
-        evmConnectors.unshift(gnosisSafeConnector)
+        if (!wallet) return
 
-        connect(gnosisSafeConnector)
-        return
-      }
-      if ((!source || !type) && !isGnosis) return
-      if (type === 'substrate') {
+        if (wallet.connector.connectEagerly) {
+          await wallet.connector.connectEagerly()
+        } else {
+          await wallet.connector.activate()
+        }
+
+        dispatch({ type: 'evmSetState', payload: { selectedWallet: wallet } })
+        dispatch({ type: 'setConnectedType', payload: 'evm' })
+      } else if (type === 'substrate' && source) {
         // This script might have loaded quicker than the wallet extension,
         // so we'll wait up to 2 seconds for it to load
         let i = 8
@@ -68,25 +50,7 @@ export function useConnectEagerly(
           }
           await new Promise((res) => setTimeout(res, 250))
         }
-      } else if (type === 'evm') {
-        const wallets = evmConnectors.filter((c) => c.id === source || c.connector instanceof GnosisSafe)
-        const connected = await Promise.allSettled(
-          wallets.map(async (wallet) => {
-            if (wallet.connector.connectEagerly) {
-              await wallet.connector.connectEagerly()
-            } else {
-              await wallet.connector.activate()
-            }
-            return wallet
-          })
-        )
-        const first = connected.find((result) => result.status === 'fulfilled') as
-          | PromiseFulfilledResult<EvmConnectorMeta>
-          | undefined
-        if (first) {
-          dispatch({ type: 'evmSetState', payload: { selectedWallet: first.value } })
-          dispatch({ type: 'setConnectedType', payload: 'evm' })
-        }
+      } else {
       }
     } finally {
       setIsTrying(false)
